[fix] new publish OK
