+++ src/middlewares/hashtagValidator.js
@@ -1,6 +1,6 @@
 import { getHashtagByName } from "../repositories/hashtagRepository.js";
 
-async function hashtagValidator(req, res, next){
+export async function hashtagValidator(req, res, next){
 
     const hashtagName = req.params.hashtag;
 
@@ -19,30 +19,4 @@ async function hashtagValidator(req, res, next){
     } catch {
         res.sendStatus(500);
     }
-}
-
-async function hashtagChecker(req, res, next){
-
-    const { content } = req.body;
-    let hashtagArray = [];
-
-    if(content){
-        const lookForHashtags = content.split(" ");
-
-        for(let i = 0; i < lookForHashtags.length; i++){
-            if(lookForHashtags[i].substr(0,1) === "#"){
-                hashtagArray.push(lookForHashtags[i].substr(1,lookForHashtags[i].length));
-            }
-        }
-    }
-
-    res.locals.hashtagArray = hashtagArray;
-    
-    next();
-
-}
-
-export {
-    hashtagValidator,
-    hashtagChecker
 }
