+++ src/repositories/publishRepository.js
@@ -1,12 +1,8 @@
 import db from "../database/postgres.js";
 
 
-function createPost(post){
+export function newPublish(post){
 
  const {userId, url, content} = post;
-    return db.query('INSERT INTO posts ("userId",content, url) VALUES ($1, $2, $3) RETURNING id', [userId, content, url])
-}
-
-export {
-    createPost
+    return db.query('INSERT INTO posts ("userId", content, url) VALUES ($1, $2, $3) RETURNING id', [userId, content, url])
 }
